Zodify LanguageCode, move dynamo types to common types library
